\name{BI_ouput}
\alias{BI_output}
\docType{data}
\title{Output data for Box Ironbark Forest and Woodland state and transition models.}
\description{
}
\usage{data(BI_output)}
\format{
  \describe{
    \item{\code{}}{}
  }
}
\details{
 The \code{BI_output} data are described in Czembor etal (2011).

\source{

}
\references{}
\examples{
data(BI_output)
str(BI_output)
}
\keyword{datasets}
