% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifaw.r
\docType{data}
\name{bifaw}
\alias{bifaw}
\title{Box Ironbark Forest and Woodland Data}
\format{A list with two elements
\describe{
  \item{inputs}{input data}
  \item{ouputs}{model outputs}
  ...
}}
\usage{
bifaw
}
\description{
A dataset containing the expert elicited state and transition model inputs and outputs
}
\keyword{datasets}

