% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoi.r
\name{evoi}
\alias{evoi}
\title{Expected Value of Information}
\usage{
evoi(data, ncores = 1, max_eco = 0.2, type = c("evpi", "evwoi",
  "evpxi_list", "evpxi"), parameter_list = NULL)
}
\arguments{
\item{data}{Input and output data as a single data.frame.}

\item{ncores}{Interger. Number of cpu cores.}

\item{max_eco}{Numeric. Maximum proportion of the landscape allocatable to 
thinning.}

\item{type}{Character. Type of evoi analysis.}

\item{parameter_list}{list. When \code{type = "evpxi_list"} a list of
parameter sets must be supplied where the list elements are named
vectors of parameter where the names correspond to the action the
parameter of interest associated with.}
}
\description{
Calculate the expected value of information for box ironbark and woodland 
management
}

