\name{pre_posterior}

\alias{pre_posterior}

\title{Update a prior distribution.}

\description{Create a set of preposterior distributions by updating a unit interval prior distribution with hypothetical binomial trials.}

\usage{}

\arguments{
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
}

\details{}

\value{

}

\references{

}

\author{
William K. Morris.
}

\seealso{
}

\examples{
}
