\name{sensitivity_BI}

\alias{sensitivity_BI}
\alias{sens_prepost_BI}

\title{Sensitivity analyses of the Box Ironbark state and transition model.}

\description{Perform sensitivity analyses for the Box Ironbark state and transition models either over the range of the }

\usage{}

\arguments{
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
 \item{}{}
}

\details{}

\value{

}

\references{

}

\author{
William K. Morris.
}

\seealso{
}

\examples{
}
